content-pipeline/requirements.txt
moviepy==2.1.2
Pillow>=10.0.0
edge-tts>=6.1.0
google-generativeai>=0.8.0
groq>=0.11.0
huggingface-hub>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
